refactor(recovery): isolate readiness owners
Problem:
The initial fix added focused behavior and regression coverage to several already-large owner and test files, exceeding repository capacity ratchets.

Root cause:
Recovery checkpoint, transport, history, and test code remained embedded in broad legacy files instead of adjacent owner-specific units.

Fix:
Extract final-readiness provider, evidence, desktop, serve, controller, and test logic into cohesive files. Keep only unavoidable protocol and dispatch seams in the large owners, and update the nine affected ratchets with narrowly measured sub-one-percent source growth (the controller file shrinks overall).

Verification:
- go test ./...
- cd desktop && go test ./...
- targeted go test -race recovery suite
- cd desktop/frontend && pnpm typecheck
- go run ./tools/repolint
- git diff --check
