+++ internal/agent/run_loop.go
@@ -120,17 +120,8 @@ func (a *Agent) beginRunTurn(ctx context.Context, input string) (rawInput string
 	a.turn = turnRuntime{}
 	a.resetStructuralRunGuards()
 	scope, scoped := DeliveryExecutionScopeFromContext(ctx)
-	preserveEvidence := a.pending.preserveEvidence
-	// A pending card belongs only to the immediately preceding turn. Starting
-	// any new user turn consumes its durable action; explicit recovery has
-	// already restored the checkpoint above via preserveEvidence.
-	if a.sess.conversation != nil {
-		a.sess.conversation.ConsumeFinalReadinessRecovery()
-	}
-	// A run that starts with a pending readiness recovery (or an explicit
-	// evidence-preserving continuation) and then passes readiness counts as a
-	// recovery in the final audit.
-	a.turn.readinessRecovered = preserveEvidence || a.pending.finalReadinessRecovery
+	preserveEvidence, readinessRecovered := a.beginFinalReadinessRecovery()
+	a.turn.readinessRecovered = readinessRecovered
 	if a.task.ledger != nil {
 		switch {
 		case preserveEvidence:
@@ -140,11 +131,6 @@ func (a *Agent) beginRunTurn(ctx context.Context, input string) (rawInput string
 		default:
 			a.resetTurnEvidence()
 		}
-	}
-	a.pending.preserveEvidence = false
-	a.pending.finalReadinessRecoveryPrepared = false
-	if !preserveEvidence {
-		a.pending.finalReadinessRecovery = false
 	}
 	if scoped {
 		a.task.scopeID = scope.ID
