+++ internal/agent/turnruntime.go
@@ -108,10 +108,8 @@ type pendingTurn struct {
 	// of resetting it, so a review_report completion nudge can cite the read
 	// receipts the subagent already earned. Consumed by that Run.
 	preserveEvidence bool
-	// finalReadinessRecovery is armed only when this agent exhausts final
-	// readiness.
-	// An explicit host recovery action can consume it to preserve the failed
-	// turn's receipts once; an ordinary user turn still resets evidence.
+	// finalReadinessRecovery is armed after final readiness fails. An explicit
+	// host action preserves receipts once; an ordinary turn resets evidence.
 	finalReadinessRecovery bool
 	// finalReadinessRecoveryPrepared prevents the durable marker fallback from
 	// being consumed twice before the prepared Run starts.
