+++ internal/control/controller.go
@@ -1107,17 +1107,6 @@ func (c *Controller) RunTurn(ctx context.Context, input string) error {
 	})
 }
 
-// RunFinalReadinessRecovery is the synchronous transport-neutral recovery path
-// used by ACP and other request/response frontends.
-func (c *Controller) RunFinalReadinessRecovery(ctx context.Context, input string) error {
-	return c.runSynchronousTurn(ctx, nil, func(runCtx context.Context) error {
-		if c.executor == nil || !c.executor.PrepareFinalReadinessRecovery() {
-			return ErrNoFinalReadinessRecovery
-		}
-		return c.runTurn(runCtx, input)
-	})
-}
-
 func (c *Controller) runTurnWithRaw(ctx context.Context, input, raw string) error {
 	return c.runTurnWithRawDisplay(ctx, input, raw, "")
 }
@@ -1250,26 +1239,6 @@ func isNonTurnHTTPInput(input string) bool {
 // text for transcript replay when controller-side composition expands input.
 func (c *Controller) SubmitDisplay(display, input string) {
 	c.submit(input, display, "")
-}
-
-// SubmitFinalReadinessRecovery runs the same visible prompt path as
-// SubmitDisplay but first authorizes the executor to retain the immediately
-// preceding exhausted ledger. The authorization is durable across controller
-// rebuilds and one-shot; stale actions fail instead of inheriting unrelated
-// evidence or silently becoming ordinary turns.
-func (c *Controller) SubmitFinalReadinessRecovery(display, input string) {
-	c.runGuarded(func(ctx context.Context) error {
-		if c.executor == nil || !c.executor.PrepareFinalReadinessRecovery() {
-			return ErrNoFinalReadinessRecovery
-		}
-		return c.runGoalLoopWithRawDisplay(ctx, input, input, display)
-	})
-}
-
-// SubmitDeliveryRecovery preserves the v1.25 desktop/API symbol while routing
-// through the generic readiness owner.
-func (c *Controller) SubmitDeliveryRecovery(display, input string) {
-	c.SubmitFinalReadinessRecovery(display, input)
 }
 
 // SubmitInvocationDisplay executes composer-selected invocation entities
@@ -1450,14 +1419,10 @@ func (c *Controller) submitCommandOrTurnReady(trimmed, input, display string, sc
 			return c.runEditedGoalLoopWithRawDisplay(ctx, input, raw, display, editedOriginal)
 		}
 	}
+	if c.submitFinalReadinessCommand(trimmed, display) {
+		return
+	}
 	switch {
-	case trimmed == ContinueChecksCommand || strings.HasPrefix(trimmed, ContinueChecksCommand+" "):
-		prompt, _ := ParseFinalReadinessRecoveryCommand(trimmed)
-		shown := display
-		if strings.TrimSpace(shown) == "" {
-			shown = trimmed
-		}
-		c.SubmitFinalReadinessRecovery(shown, prompt)
 	case trimmed == "/compact" || strings.HasPrefix(trimmed, "/compact "):
 		focus := strings.TrimSpace(strings.TrimPrefix(trimmed, "/compact"))
 		go func() {
